requirements.txt
aiohttp==3.6.1
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sorl-thumbnail==12.5.0
sqlparse==0.3.0
yarl==1.3.0
